+++ src/TokenizerFwf.h
@@ -13,14 +13,15 @@ class TokenizerFwf : public Tokenizer {
   SourceIterator begin_, cur_, curLine_, end_;
   int row_, col_, cols_, max_;
   std::string comment_;
-  bool moreTokens_, isRagged_, hasComment_;
+  bool moreTokens_, isRagged_, hasComment_, trimWS_;
 
 public:
   TokenizerFwf(
       const std::vector<int>& beginOffset,
       const std::vector<int>& endOffset,
       std::vector<std::string> NA = std::vector<std::string>(1, "NA"),
-      std::string comment = "");
+      std::string comment = "",
+      bool trimWS = true);
 
   void tokenize(SourceIterator begin, SourceIterator end);