+++ tests/testthat/test-read-fwf.R
@@ -9,6 +9,29 @@ test_that("trailing spaces ommitted", {
   expect_equal(df$X1, df$X2)
 })
 
+test_that("respects the trim_ws argument", {
+  x <- "a11 b22 c33\nd   e   f  "
+  out1 <- read_fwf(x, fwf_empty(x), trim_ws = FALSE)
+  expect_equal(out1$X1, c("a11", "d  "))
+  expect_equal(out1$X2, c("b22", "e  "))
+  expect_equal(out1$X3, c("c33", "f  "))
+
+  out2 <- read_fwf(x, fwf_empty(x), trim_ws = TRUE)
+  expect_equal(out2$X1, c("a11", "d"))
+  expect_equal(out2$X2, c("b22", "e"))
+  expect_equal(out2$X3, c("c33", "f"))
+})
+
+test_that("respects the trim_ws argument with empty fields", {
+  x <- "a11 b22 c33\nd       f  "
+  out1 <- read_fwf(x, fwf_empty(x), trim_ws = FALSE)
+  expect_equal(out1$X1, c("a11", "d  "))
+  expect_equal(out1$X2, c("b22", "   "))
+  expect_equal(out1$X3, c("c33", "f  "))
+
+  out1 <- read_fwf(x, fwf_empty(x), trim_ws = TRUE, na = "NA")
+})
+
 test_that("skipping column doesn't pad col_names", {
   x <- "1 2 3\n4 5 6"